Changed uint to size_t to compile on x64. Don't know the code or protocol, so this might not be correct

+++ sha1.d
@@ -203,7 +203,7 @@ public:
       return rv;
    }
 
-   bool input(const(ubyte)* message_array, uint length)
+   bool input(const(ubyte)* message_array, size_t length)
    {
       if (!length)
          return true;
